Revert "added 3*3 mouse bites between the sub-boards on both edges"
This reverts commit 72bb31031476af87d888838294cc72f0395ed406.

it turns out that mouse bites shorts the rails if the board house does
not adhere to the NPTH drill directions, which kind of makes up for the
usefulness that the mouse bites provided. and they did not improve the
board sturdiness significantly, so it is not a problem to remove them.